pre-release demo

+++ sketchy/terminal/predict/commands.py
@@ -14,6 +14,9 @@ from sketchy.sketchy import Sketchy
     '--sketch', '-s', default='saureus-v1.msh', help='MASH sketch to query'
 )
 @click.option(
+    '--data', '-d', help='Index data file to pull genotypes and other data from.'
+)
+@click.option(
     '--tmp', '-t', default=Path().home() / '.sketchy' / 'tmp',
     help='Temporary dir for slicing Fastq file'
 )
@@ -26,24 +29,35 @@ from sketchy.sketchy import Sketchy
 @click.option(
     '--reads', '-r', default=50, help='Number of reads to type.'
 )
-def predict(fastq, sketch, tmp, cores, header, reads):
+def predict(fastq, sketch, data, tmp, cores, header, reads):
 
-    """ Online lineage matching and antibiotic susceptibility predictions using
-    MinHash from uncorrected nmanopore reads"""
+    """ Online lineage matching from uncorrected nanopore reads"""
 
     sketchy = Sketchy()
 
+    fastq_path = Path(fastq).resolve()
+    sketch_path = Path(sketch).resolve()
+
+    if not fastq_path.exists():
+        print(f'File {fastq_path} does not exist.')
+        exit(1)
+
+    if not sketch_path.exists():
+        print(f'Mash sketch {sketch_path} does not exist.')
+        exit(1)
+
     try:
         sketchy.predict_nanopore(
-            sketch=Path(sketch).resolve(),
-            fastq=Path(fastq).resolve(),
+            sketch=sketch_path,
+            fastq=fastq_path,
             tmp=Path(tmp).resolve(),
             cores=cores,
             header=header,
             nreads=reads,
+            data=data,
         )
 
     except KeyboardInterrupt:
         shutil.rmtree(tmp)
-    finally:
+    else:
         shutil.rmtree(tmp)
